[Product] 안전 재고 수량 타입 BIGINT로 변경
- ERD 결정 사항(수량 컬럼은 모두 BIGINT)에 맞춰 product_sku.safety_stock_quantity를 DECIMAL(14,3)에서 BIGINT로 변경
- 적용된 V2 체크섬 보호를 위해 V5 마이그레이션으로 분리
- ProductSku 도메인, JPA 엔티티, 등록 요청/커맨드, 응답 DTO, 테스트의 타입을 BigDecimal에서 Long으로 변경

+++ src/main/java/com/kb/wms/product/adapter/in/web/dto/response/ProductSkuCreateResponse.java
@@ -19,7 +19,7 @@ public record ProductSkuCreateResponse(
         BigDecimal currentPurchasePrice,
         BigDecimal currentSupplyPrice,
         String unit,
-        BigDecimal safetyStockQuantity,
+        Long safetyStockQuantity,
         boolean isActive,
         List<SkuOptionValueResponse> optionValues,
         LocalDateTime createdAt
